pyproject.toml
[tool.poetry]
name = "nonebot-plugin-chatgpt-turbo"
version = "1.2.0"
description = "A nonebot plugin for oneapi and openai"
authors = ["Alpaca <alpaca@bupt.edu.cn>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = "^2.2.1"
openai = "^1.55.3"
nonebot_plugin_saa = "^0.7.1"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
